requirements.txt
matplotlib==3.2.1
mtcnn==0.1.0
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
